requirements.txt
Bootstrap-Flask==1.8.0
click==8.0.3
colorama==0.4.4
dependency-injector==4.36.2
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pymongo==3.12.0
python-dotenv==0.19.1
six==1.16.0
Werkzeug==2.0.2

numpy~=1.21.3
requests~=2.26.0
pydantic~=1.8.2
pyspellchecker